refactor: html to js (#5838)
* refactor: html to js

* fix: config

* revert: keep renderToEntry

* fix: types

* refactor: simplify dist type

* chore: changelog
