+++ packages/plugin-pha/tests/mockServer.mjs
@@ -1,4 +1,4 @@
-export function renderToEntry(requestContext, documentOnly) {
+export function renderToHTML(requestContext, documentOnly) {
   const { req } = requestContext;
   return {
     value: `<html><body>${req.url}${documentOnly ? '-document' : ''}</body></html>`,
