+++ packages/runtime/src/index.server.ts
@@ -1,2 +1,2 @@
-export { renderToEntry, renderToResponse, renderToHTML } from './runServerApp.js';
+export { renderToResponse, renderToHTML, renderToEntry } from './runServerApp.js';
 export * from './index.js';
