+++ packages/runtime/src/runServerApp.tsx
@@ -1,12 +1,7 @@
 import type { ServerResponse } from 'http';
-import * as path from 'path';
-import { fileURLToPath } from 'url';
 import * as React from 'react';
 import * as ReactDOMServer from 'react-dom/server';
 import { Action, parsePath } from 'history';
-import * as htmlparser2 from 'htmlparser2';
-import ejs from 'ejs';
-import fse from 'fs-extra';
 import type { Location } from 'history';
 import type {
   AppContext, RouteItem, ServerContext,
@@ -17,7 +12,6 @@ import type {
   DocumentComponent,
   RuntimeModules,
   AppData,
-  DistType,
 } from './types.js';
 import Runtime from './runtime.js';
 import App from './App.js';
@@ -33,14 +27,7 @@ import getRequestContext from './requestContext.js';
 import matchRoutes from './matchRoutes.js';
 import getCurrentRoutePath from './utils/getCurrentRoutePath.js';
 import DefaultAppRouter from './AppRouter.js';
-import __createElement from './domRender.js';
-
-let dirname;
-if (typeof __dirname === 'string') {
-  dirname = __dirname;
-} else {
-  dirname = path.dirname(fileURLToPath(import.meta.url));
-}
+import { renderHTMLToJS } from './renderHTMLToJS.js';
 
 interface RenderOptions {
   app: AppExport;
@@ -60,7 +47,7 @@ interface RenderOptions {
     [key: string]: PageConfig;
   };
   runtimeOptions?: Record<string, any>;
-  distType?: Array<'html' | 'javascript'> | 'html' | 'javascript';
+  distType?: Array<'html' | 'javascript'>;
 }
 
 interface Piper {
@@ -70,24 +57,35 @@ interface Piper {
 interface RenderResult {
   statusCode?: number;
   value?: string | Piper;
-  jsOutput?: string;
+}
+
+/**
+ * Render and send the result with both entry bundle and html.
+ */
+export async function renderToEntry(
+  requestContext: ServerContext,
+  renderOptions: RenderOptions,
+) {
+  const result = await renderToHTML(requestContext, renderOptions);
+  const { value } = result;
+
+  let jsOutput;
+
+  const { distType } = renderOptions;
+  if (value && distType && distType.includes('javascript')) {
+    jsOutput = await renderHTMLToJS(value);
+  }
+
+  return {
+    ...result,
+    jsOutput,
+  };
 }
 
 /**
  * Render and return the result as html string.
  */
 export async function renderToHTML(
-  requestContext: ServerContext,
-  renderOptions: RenderOptions,
-): Promise<RenderResult> {
-  renderOptions.distType = 'html';
-  return renderToEntry(requestContext, renderOptions);
-}
-
-/**
- * Render and return the result as entry string.
- */
-export async function renderToEntry(
   requestContext: ServerContext,
   renderOptions: RenderOptions,
 ): Promise<RenderResult> {
@@ -124,13 +122,12 @@ export async function renderToEntry(
  */
 export async function renderToResponse(requestContext: ServerContext, renderOptions: RenderOptions) {
   const { res } = requestContext;
-  const { distType } = renderOptions;
   const result = await doRender(requestContext, renderOptions);
 
   const { value } = result;
 
   if (typeof value === 'string') {
-    sendResult(res, result, distType);
+    sendResult(res, result);
   } else {
     const { pipe, fallback } = value;
 
@@ -147,7 +144,7 @@ export async function renderToResponse(requestContext: ServerContext, renderOpti
         // downgrade to CSR.
         console.error('PipeToResponse onShellError, downgrade to CSR.', err);
         const result = await fallback();
-        sendResult(res, result, distType);
+        sendResult(res, result);
       },
       onError: async (err) => {
         // onError triggered after shell ready, should not downgrade to csr.
@@ -160,15 +157,10 @@ export async function renderToResponse(requestContext: ServerContext, renderOpti
 /**
  * Send string result to ServerResponse.
  */
-async function sendResult(res: ServerResponse, result: RenderResult, distType: DistType) {
+async function sendResult(res: ServerResponse, result: RenderResult) {
   res.statusCode = result.statusCode;
-  if (distType && distType.includes('javascript') || distType === 'javascript') {
-    res.setHeader('Content-Type', 'text/js; charset=utf-8');
-    res.end(result.jsOutput);
-  } else {
-    res.setHeader('Content-Type', 'text/html; charset=utf-8');
-    res.end(result.value);
-  }
+  res.setHeader('Content-Type', 'text/html; charset=utf-8');
+  res.end(result.value);
 }
 
 async function doRender(serverContext: ServerContext, renderOptions: RenderOptions): Promise<RenderResult> {
@@ -184,7 +176,6 @@ async function doRender(serverContext: ServerContext, renderOptions: RenderOptio
     runtimeModules,
     renderMode,
     runtimeOptions,
-    distType,
   } = renderOptions;
   const finalBasename = serverOnlyBasename || basename;
   const location = getLocation(req.url);
@@ -226,9 +217,6 @@ async function doRender(serverContext: ServerContext, renderOptions: RenderOptio
   }
 
   const matches = matchRoutes(routes, location, finalBasename);
-  if (distType === 'javascript' || (Array.isArray(distType) && distType.includes('javascript'))) {
-    return renderDocument({ matches, renderOptions });
-  }
 
   const routePath = getCurrentRoutePath(matches);
   if (documentOnly) {
@@ -337,69 +325,6 @@ interface RenderDocumentOptions {
   renderOptions: RenderOptions;
   routePath?: string;
   downgrade?: boolean;
-  distType?: Array<'html' | 'javascript'>;
-}
-
-function renderDocumentToJs(html) {
-  let jsOutput = '';
-  const dom = htmlparser2.parseDocument(html);
-
-  let headElement;
-  let bodyElement;
-  function findElement(node) {
-    if (headElement && bodyElement) return;
-
-    if (node.name === 'head') {
-      headElement = node;
-    } else if (node.name === 'body') {
-      bodyElement = node;
-    }
-
-    const {
-      children = [],
-    } = node;
-    children.forEach(findElement);
-  }
-  findElement(dom);
-
-  const extraScript = [];
-  function parse(node) {
-    const {
-      name,
-      attribs,
-      data,
-      children,
-    } = node;
-    let resChildren = [];
-
-    if (children) {
-      if (name === 'script' && children[0] && children[0].data) {
-        extraScript.push(`(function(){${children[0].data}})();`);
-      } else {
-        resChildren = node.children.map(parse);
-      }
-    }
-
-    return {
-      tagName: name,
-      attributes: attribs,
-      children: resChildren,
-      text: data,
-    };
-  }
-
-  const head = parse(headElement);
-  const body = parse(bodyElement);
-
-  const templateContent = fse.readFileSync(path.join(dirname, '../templates/js-entry.js.ejs'), 'utf-8');
-  jsOutput = ejs.render(templateContent, {
-    createElement: __createElement,
-    head,
-    body,
-    extraScript,
-  });
-
-  return jsOutput;
 }
 
 /**
@@ -420,7 +345,6 @@ function renderDocument(options: RenderDocumentOptions): RenderResult {
     Document,
     basename,
     routesConfig = {},
-    distType = ['html'],
   } = renderOptions;
 
   const routesData = null;
@@ -462,14 +386,8 @@ function renderDocument(options: RenderDocumentOptions): RenderResult {
     </AppContextProvider>,
   );
 
-  let jsOutput = '';
-  if (distType.includes('javascript') || distType === 'javascript') {
-    jsOutput = renderDocumentToJs(htmlStr);
-  }
-
   return {
     value: `<!DOCTYPE html>${htmlStr}`,
-    jsOutput,
     statusCode: 200,
   };
 }
